[DEV] Combined controller utilities and resource utilites together.
