+++ Tests/DebouncerTests.java
@@ -1,7 +1,4 @@
-import controllers.utilities.Debouncer;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
+import utilities.Debouncer;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
