+++ src/main/java/utilities/Utilities.java
@@ -1,4 +1,4 @@
-package controllers.utilities;
+package utilities;
 
 import commands.Command;
 import commands.SwitchToDarkMode;
@@ -8,7 +8,9 @@ import data.Storage;
 import data.database.Database;
 import views.*;
 
+import java.net.URL;
 import java.util.List;
+import java.util.Objects;
 
 /*
  This class contains utility methods used amongst the controllers.
@@ -49,4 +51,32 @@ public class Utilities {
         }
     }
 
+    /**
+     * Load a file from the resource folder in the format of
+     * a URL.
+     *
+     * @param path Path to the file relative to the resource folder.
+     */
+    public static URL loadFileByURL(String path) {
+        return Objects.requireNonNull(
+                Utilities.class.getClassLoader()
+                        .getResource(path)
+        );
+    }
+
+    /**
+     * Return true if a string is an integer.
+     *
+     * @param s The string.
+     * @return True if it is a integer, false otherwise.
+     */
+    public static boolean isInteger(String s) {
+        try {
+            Integer.parseInt(s);
+        } catch (NullPointerException | NumberFormatException e) {
+            return false;
+        }
+
+        return true;
+    }
 }
